+++ mod/debugadapter.lua
@@ -8,6 +8,13 @@ if __Profiler then
   return
 end
 
+if data then
+  -- data stage clears package.loaded between files, so we stash a copy in Lua registry too
+  local reg = debug.getregistry()
+  local regDA = reg.__DebugAdapter
+  if regDA then return regDA end
+end
+
 -- this is a global so the vscode extension can get to it from debug.debug()
 __DebugAdapter = __DebugAdapter or {} -- but might have been defined already for selective instrument mode
 local __DebugAdapter = __DebugAdapter
@@ -298,6 +305,9 @@ do
     __DebugAdapter.attach()
     print("DBG: on_data")
     debug.debug()
+    -- data stage clears package.loaded between files, so we stash a copy in Lua registry too
+    local reg = debug.getregistry()
+    reg.__DebugAdapter = __DebugAdapter
   elseif script.mod_name ~= "debugadapter" then -- don't hook myself!
     -- in addition to the global, set up a remote so we can configure from DA's on_tick
     -- and pass stepping state around remote calls